requirements.txt
Flask 
Pillow 
pytesseract 
pyproj 
folium
streamlit
streamlit-folium